requirements.txt
python-pptx
pandas
numpy
seaborn
matplotlib
openpyxl